Moves `FieldDefinition` from `dto.factory.types` to `dto.factory.data_structures`.
