+++ litestar/contrib/sqlalchemy/dto.py
@@ -17,8 +17,8 @@ from sqlalchemy.orm import (
 )
 
 from litestar.dto.factory.abc import AbstractDTOFactory
+from litestar.dto.factory.data_structures import FieldDefinition
 from litestar.dto.factory.field import DTO_FIELD_META_KEY, DTOField, Mark
-from litestar.dto.factory.types import FieldDefinition
 from litestar.dto.factory.utils import get_model_type_hints
 from litestar.types.empty import Empty
 from litestar.utils.helpers import get_fully_qualified_class_name
